+++ src/UI/Core/TextArea.h
@@ -14,7 +14,7 @@ public:
 	TextArea() = default;
 	virtual ~TextArea() = default;
 
-	void textColor(int r, int g, int b, int a = 255) { mTextColor(r, g, b, a); }
+	void textColor(uint8_t r, uint8_t g, uint8_t b, uint8_t a = 255) { mTextColor(r, g, b, a); }
 
 	void font(const std::string&, size_t);
